Subtract UCB1 exploration term when minimizing
Bug fix: because lower UCB values are better when minimizing an
objective, we should subtract the exploration term, to make an
under-explored arm have a lower UCB value as the total number
of iterations increases. This fixes CUR-1550, where some arms were
getting stuck with a handful of visits and then never being
explored again.